backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
transformers
pymongo[srv]
pydantic
pydantic-settings
dnspython
torch
scipy
google-generativeai
edge-tts  